refactor kubelet/server changes into a small patch; refactor kubelet cmd design to align with upstream impl
